+++ Changelog.md
@@ -1,3 +1,3 @@
 ## v0.0.1
 
- * Supports export of database scheme to a TYPO3 Extbase Extension using Symfony database connections.
+ * Supports export of database schema to a TYPO3 Extbase Extension using Symfony database connections.